requirements.txt
mcp>=1.3.0
httpx>=0.23.0
pydantic>=1.10.0
aiohttp>=3.8.4
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
